third_party/WebKit/LayoutTests/fast/dom/shadow/new-style-inserted-into-parent-treescope-expected.txt

Test the case: a treescope has some style element and a new style is inserted into its parent treescope.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(getNodeInTreeOfTrees('host1/host2/target')).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
